WIP: Fix query 'buildstarttime' start and end datetimes based on testing day start time (#288)
This change correctly sets the start and begin 'buildstartime' filter fields
for test and build history to match the CDash project's testing day start time.

An important thing this adds is that you nave to explicitly set the timezone
for the 'buildstarttime' field because if you don't give the timezone, CDash
will interpret this according to the CDash server's configured timezone (same
for all CDash projects on the CDash server).  The default CDash timezone is
UTC but Sierra's CDash site this this set as:

  date_default_timezone_set ( "America/Denver" );

in the config.local.php file.  This in causes that CDash site to print
'buildstarttime' with 'MDT' and it interprets "YYYY-MM-DDThh:mm:ss" with no
explicit timezone as "MDT".  Therefore, to get the queries created correctly,
one must explicitly list the timezone 'UTC' for the 'buildstarttime' fields.

NOTE: I belive that "America/Dever" will be "MDT" in the summer and "MST" in
the winter so this is pretty confusing.

NOTE: The timezone set in the project-specific "Nightly Start Time" field set
from the CDash project confiugration page is not related and has no impact at
all on the timzone set using date_default_timezone_set() in the
config.local.php file.  This is super confusing.  We will need to document
this better once we get this fully worked out.