Add tests and docuemntation for <Project>_ENABLE_INSTALLATION_TESTING (trilinos/Trilinos#5027)
The feature worked and did not need to be fixed.  It just needed to be tested
(so it will survive future refactorings) and documented.

If you comment out TribitsExProj_ENABLE_INSTALLATION_TESTING and
TribitsExProj_INSTALLATION_DIR in the new automated test, the build will fail
because a bunch of the lib source files have been deleted.