Add test case and fix handling of global file triggering enable of all packages
I noticed a defect in the code manifisted in the Trilinos post-push CI build
where if a global file was changed giving the special package called
"ALL_PACKAGES", it was not being enabled.  This defect must have snuck in when
I added support for the all-at-once appraoch and I did not have an automated
test for this use case.

Now there are strong automated tests for this use case for if ST or only PT
packages are enabled.  It seems to be doing the right thing.

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=360,notpassed=0 (1.38 min)
1) SERIAL_RELEASE => passed: passed=360,notpassed=0 (1.45 min)