Remove duplicate entires from test history
This will now report the correct number of test history entries even if CDash
returns duplicate tests from queryTests.php.

We were already addressing duplicate test entires when compiling the list of
non-passing tests.  This adds that same functionality to the test history
list.

NOTE: The CDash query will still show duplicate tests if the user clicks on
the link but there is not much we can do about that (other than ask Kitware to
fix this).

This has strong unit tests and I also manually tested this with ROL and ATDM
Trilinos builds.