cdash_analyze_and_report.py: Add --require-test-history-match-nonpassing-tests (#301)
This adds the option --require-test-history-match-nonpassing-tests with the
default 'off'.  I added a unit test for the use case that I need in ATDM
Trilinos testing which is to filter out a ton of known random system failures
(see trilinos/Trilinos#6861).

I also added a function getIdxOfTestInTestLOD() to improve maintenance of the
automated tests to make them less fragile to the test data.

I think this is enough testing for now.