Fix ST_LibUsage tests for CMake 3.17
For some reason, these CMake 3.17 reports an export error in one of the
libraries in these tests for some of the configure cases.  But since these
tests are checking TriBITS behavior and not so much CMake behavior, we can
disable the generation of export makefiles and not damange the strength of the
tests.