Always read in CTest configuration files for all-at-once mode (#326)
This makes the updated test from the previous commit pass and proves that this
use case is now fixed (i.e. not running configure but running tests).

NOTE: The fact that all of the tests still pass, even the ones where the
configure fails and does not produce a file CTestCustom.cmake.  So hopefully
this will not break any existing use cases for this software.