Don't run any extra commands if group/perms options not set (trilinos/Trilinos#7881)
Previously, the command 'stat -c %U <path>' was always being run and that
command failed on MacOSX (that uses 'stat -f $U <path>' instead as fixed in
the previous commit).  The issue is that such commands should not even be run
unless the user sets one of the group and/or permissions fixup options that
TriBITS now supports.

With this commit, when the user has not selected any of the group and/or
permissions fixup options, then TriBITS will not exectute any extra code
related to this at all durning configure, build, or install time.

A test was updated to check that the file
set_installed_group_and_permissions.cmake is not even generated in the build
dir if none of the group/perms options are set.  If you change the production
code to always generate the file set_installed_group_and_permissions.cmake,
then the test will fail (which I verified in TDD process).  If that file does
not get generated then it can't be called durring 'install' and therefore
can't call any extra (less portable) comamnds.