Add test cases for TPL_ENABLE_MPI=ON and other test name modifications (#173, trilinos/Trilinos#7112)
Turns out that TRIBITS_ADD_TEST() had bad logic for reading in
<fullTestName>_SET_RUN_SERIAL and <fullTestName>_SET_DISABLED_AND_MSG
when the test name is modifed by setting TPL_ENABLE_MPI=ON or passing in
NAME_POSTFIX or POSTFIX_AND_ARGS_<IDX> arguments.

These are obvious use cases that I missed and they are needed for
trilinos/Trilinos#7112 and in general, obviously.

This means that <fullTestName>_SET_DISABLED_AND_MSG was not implemented
correctly for TRIBITS_ADD_TEST() as part of #173.  Turns out the driving
use case for that was using TRIBITS_ADD_ADVACED_TEST() so we never noticed
the bug.

The next commit will fix this (which I already prototyped).