Fix install perms mods and tests to work on MacOSX for real (trilinos/Trilinos#7881)
In my last set of commits, I was not not able to actaully test this on a
MacOSX machine and it needed a few tweaks.  I also updated the automated tests
for this in TriBITS to 100% pass on MacOSX.  Therefore, I am pretty sure this
is doing what it is supposed to do.

The changes that I had to make were:

* The chmod command on MacOSX requires the '-R' argument to be before the
  permissions change string.  (Hopefully Linux/Bash chmod will allow the same
  also.)

* The 'ls' command on MaxOSX formats the ouput a little differently than on
  Linux/Bash and therefore small changes in the regex matching strings had to
  be made.  (This should work on Linux/Bash as well.)