Generate error when trying to set install perms/group mod options on Windows (#314)
This replaces logic that just ignored the instal perm/group modification
options for Windows platforms added as part of Story #314.  Now if the user
tries to set any of these options, it errors out with a (hopefully) good error
message.

This was flagged as part of the review of PR #327.