Update TRIBITS_ADD_TEST() to not add test when NUM_MPI_PROCS > 1 for TPL_ENABLE_MPI=OFF (trilinos/Trilinos#8110)
This avoids adding a bunch of Trilinos tests that got defined like:

  TRIBITS_ADD_TEST(
    tif_belos
    NAME Jacobi_hb_belos
    ARGS "--xml_file=test_2_Jacobi_nos1_hb.xml"
    COMM serial mpi NUM_MPI_PROCS 2
    STANDARD_PASS_OUTPUT
  )

Well, a serial test with NUM_MPI_PROCS=2 MPI processes makes no sense at all.
Therefore, we need to just not add a test like this in a serial
TPL_ENABLE_MPI=OFF build.  Before TriBITS got updated to use add_test(NAME
... COMMAND ...), these extra tests were not getting added at all.  So this
does the right thing for Trilinos (and likely every TriBITS project).  I added
a unit test to expose the problem and then fixed the code to make it pass.
This makes the serial TPL_ENABLE_MPI=OFF configuration of Trilinos work again.

To make the serial TPL_ENABLE_MPI=OFF unit testing more realistic, I moved all
of the vars that would get set in TRIBITS_SETUP_MPI() to below so the are
empty.  This is the env for how serial tests are actually defined in TriBITS.
This should be better coverage.

I also made TRIBITS_ADD_TEST() robust to the user setting
MPI_EXEC_DEFAULT_NUMPROCS > 1 when doing a serial TPL_ENABLE_MPI=OFF build.
In this case, that value gets ignored when determining the number of MPI
processes to use and is valid or not.

I also kept the behavior where setting MPI_EXEC_MAX_NUMPROCS in a serial
TPL_ENABLE_MPI=OFF build will be used to exclude tests based on the passed-in
NUM_TOTAL_CORES_USED.  That was the intention.

All of the existing unit tests pass with no change.

I updated the documentation for this as well and noted the change in behavior
for non-MPI tests.