Port to Python 3.5.2 (3#09)
This completes the port of TriBITS to Python 3.5.2 and keeps it working with
Python 2.7.5.  All of the TriBITS tests pass with both of these Python
versions!

The main thing I did here was to have runSysCmndInteface() and getCmndOutput()
return a Python unicode string for Python 3 (which is the native Python 3
string time).  This fixed a ton of code.

As part of this, I removed some convestions that Bill Spotz did years ago when
he ported TriBITS to Python 3.