Fix a bad emacs paragraph formatting change (#309)
When you use 'Ctr+Q' to reformat a paragraph Emacs, if the beginning line(s)
start with '*' it will put '*' in front of all of the following lines
(thinking you want a list of bullets).  This is very irritating.

This now builds clean with docutils (and before it did not).