Remove 's()' wrapping calls to 'getCmndOutput()' (#309)
During my refactorings for #309, I had an intermediate step where I was
wrapping calls to getCmndOutput() in the Python2and3.s() function.  But once I
refactored getCmndOutput() (and calling functions) to directly return a
Python3-native unicode string, these become unnecessary. (Kind of irratating
that the subprocess streams return byte string objects instead of native
unicode objects but so be it.)

This still passes with Python 2.7.5 and Python 3.5.2.