Fix install_package_by_package behavior for new CMP0082 behavior with CMake 3.17 upgrade (#299)
This fixes the behavior of the custom 'install_package_by_package' to
correctly install the project-level files <Project>Config.cmake and
Makefile.export.<Project> files.  To be able to invoke these install targets,
I had to move the targets into a subdir CMakeLists.txt file.  That allows
CMake -P 'cmake_install.cmake' script for just that subdir.  That way, the
install for each package can be done and the project-level installs can be
done even if one or more packages can't install their files (due to a build
error for example).

To get this to work, I had to change the binary directory for the
project-level files from CMAKE_CURRENT_BINARY_DIR to PROJECT_BINARY_DIR (which
is what they should have been all along).