Make lower-case copies of published files (SERAB-54)
There is a system at SNL that incorrectly translates URLs and makes all file
paths lower case which breaks the links to these documents.  By creating
copies that are all lower case, that fixes the problem.

I tested this manually locally and it did what it was supposed to do.