MAJOR: dashboard target: Pass build-name and 2nd CDash site vars and pass through all non-empty vars (#370)
This should resolve the problem described in #370 and make it more clear how
to control the 'dashboard' target behavior.

As part of this, I had to change the default for CTEST_DO_COVERAGE_TESTING in
TRIBITS_CTEST_DRIVER() from 'FALSE' to 'OFF' to match what gets set in the
'dashboard' target.  Running 'make dashboard' for TriBITS itself changed the
value printed out (but not the behavior of the code).  This should not be
considered a break in backward compatibility for a user but some of the
TriBITS tests are very strong and pin down the output to ensure vars are read
and used correctly.

I added a release note that this change technically breaks backward
compatibility.  See that file for more details.