Add USES_TERMINAL to 'dashboard' target (#370)
This makes it so that running the 'dashboard' target with ninja will provide
real-time STDOUT while it is running.  That is very important when using this
target (so you can monitor its progress and kill it early if it is not doing
what you want).

I noticed this while working on #370, that is all.