Increase max test STDOUT chars passed to CDash by 10x (#370)
I was noticing that output for tests like
TriBITS_CTestDriver_PBP_Dashboard_ST_ALL_PASS are getting truncated when
submitting to CDash.  I want to see the full output for all of these TriBITS
tests.

Currently this the size of the Test.xml file generated for almost the entire
TriBITS test suite is only 2.1M so size does not seem to be a problem at this
point.

I noticed this while working on #370, that is all.