Fix use case with no config running tribits_ctest_driver() (#316)
Turns out there was some uncovered code that caused a defect if there was
existing files ConfigureAttempted.txt and ConfigurePasssed.txt but you ran
with CTEST_DO_CONFIGURE=OFF.  There was a typo in a file(REMOVE ...) command.
I added a test case that triggered the defect and verified correct behavior
after the defect was fixed.