Add filename:linenum to doc blocks in rst TriBITS docuemntation (#63)
This will make it easier to find your way around the code from the developer's
guide.  This will help in the refactoring work in #63 and #299 (since it has
been years since I looked at a lot of this code).

This has the disadvantage that this makes the documentation more strongly tied
to a single version of TriBITS but the truth is that is always going to be the
case.  Customers of TriBITS that have their own copy of TriBITS should really
build the TriBITS documentation for thier version of TriBITS and post it on a
website where their developers can get to it.