Add helper functions for starting and stopping optional code timers (#63, #299)
This just cleans up the code and makes it easier to read which makes
refactorings like is needed for #63 and #299 easier.

I also changed the text for the timer for reading in the dependency files and
creating the dependency graph.  This should make this more clear what this is
doing hopefully.  (That is why I had to update one of the tests :-))