Put back INSTMACRO upper-case again (#274)
This is a case of C code being listed in a *.cmake file and getting made
lower-case by accident.

NOTE: There is not test in TriBITS for this code.  This code should not be in
TriBITS, it should be in Trilinos.  That needs to be a refactoring done in the
future to clear out TriBITS.