Remove endfunction(<string>) and endmacro(<string>) (#274, #382)
This is a follow-on from the lower-casing in #274 with PR #379.  It
lower-cased the function and macro names at the opening macro(<name> ...) and
function(<name> ...) but not the endmacro(<name> ...) and endfunction(<name>
...) calls.  (I did not realize the TriBITS had any of those still.)  Turns
out that when the text in the opening macro(<text>) and function(<text>) does
not exactly match the endmacro(<text>) and endfunction(<text>), you get a
nasty warning message:

  A logical block opening on the line ... closes on the line ... with
  mis-matching arguments.

(see #382).

The TriBITS test suite does not actually run any of this code so this was
never seen prior to the creation of #382.  Therefore, I don't know that this
solves the problem but I strongly suspect that it will.  (I will run a test
with an ATDM Trilinos build that should trigger this code.)