Add argument --skip-final-generation to build docs  (#386)
This skips final geneation of documentation.

NOTE: It would be hard to add a TriBITS test for this since the documenation
is built in-source and there is already a test that builds the documentation.
Seems like a small risk not to test this argument since this is just a
time-saving optimization when building the Sphinx documentation.