Make all builds use makefiles for now (#363)
I suspect that some of the tests are not robust enough to work with Ninja as
the outer generator.  There are some tests that will use Ninja for the inner
configure to make sure that it works.

Making the entire TriBITS test suite pass using Ninja as the generator in
place of Makefiles would be a lot of work.  Or, we could conditionally enable
many of the tests that depend Unix Makefile logic.