Disable a bunch of tests that fail if there is no Fortran (#363)
All of these tests fail to configure because we don't pass down the
ENABLE_Fortran configure option to the inner CMake configure and if a working
Fortran compiler does not even exist on the system, then these tests fail in
the configure of the inner project.  While I could pass down this info
somehow, a lot of these tests would fail anyway without some adjustment.

We don't need every test to run if we don't have a Fortran compiler.  We just
don't want the TriBITS test suite to have a bunch of failing tests if a user
clones the repo and tries to build and run the TriBITS test suite on a machine
with no Fortran compiler (like a Mac or a Windows machine).