Make WORLD_READABLE=TRUE set same perms as GROUP_WRITABLE (#314, #363)
With CMake 3.21, with umask=g-rwx,o-rwx you need explicitly set g+rX for the
group permissions to be r-X when setting
<Project>_MAKE_INSTALL_WORLD_READABLE=ON which was doing just:

  chmod -R o+rX <dir>

Now, when you ony set <Project>_MAKE_INSTALL_WORLD_READABLE=ON, you get:

  chmod -R g+rX,o+rX <dir>

For some reason, with older verisons of CMake, it was setting the default
group permission as 'r-x' when it should not have been.  I ran exactly the
same build on the same machine (crf450) with the same env except I used CMake
3.21.0 vs. 3.17.1 and I had to make this change to get the same permissions on
the base install directory.  Strange.