Add missing 'Generic' to title of generic build reference (#386)
This is not really the build reference for the TriBITS project per-say.  It
the build reference for an arbitrary TriBITS project <Project>.  It does not
make sense, I don't think, to produce this document for the TriBITS project
itself since it only has one package.  And people who are building the TriBITS
package and running TriBITS test will know quite a lot about TriBITS so I
think they can live with this generic document.