Uses build_sphinx_docs.sh in test and add rebuild (#386)
It is important to be able to rebuild the documentation iteratively locally in
a robust way.

The rebuild makes this an expensive tests by TriBITS standards (over 35s on my
super fast Linux machine) but this test is disabled by default and it only
takes one core to run so hopefully this is not too bad.