Reformat to fit in 90 char wide (#386)
I changed from 4-char to 2-char indents and wrapped lines to fit in 90 chars.
This really should be reformatted a little closer to 85 chars or so but at
least now I can do side-by-side editing on a reasonable side screen.

I know the Python standard is 4 char indents but if you do that you need a lot
more functions and a lot more wrapped lines to fit in 85 chars or so.