Factor out change_paths_and_get_includes() and is_rst_file() to be unit tested (#386)
First, the function is_rst_file() no benefit from being a member of the class.
This is not Java.  Every function does not have to belong to a class.

Second, the function change_paths_and_get_includes() has a very clear purpose
and is unit testable but not when it was inside of the class.  Removing it
from the clas will allow it to be unit tested which will allow us to find and
fix a defect quickly.