Add arg change_paths_and_get_includes( ... copy_file: bool = True ) (#386)
When unit testing, you can't copy these files.  I have no idea how this unit
test passed locally when I ran it.  In production mode, the default is to copy
the files.  This is actually a pretty nice algorithm.

I also moved the unit test higher in the CMakeLists.txt file so this might
make it more likely that this unit test runs before the actual build of the
documentation.