Add support for CTEST_UPDATE_VERSION_ONLY (#154)
Can turn this on for things like GitHub Actions and other processes that take
care of the repo versioning.  This allows you to call ctest_update() and
provide a link back to the commit on GitHub but not have ctest do any update
action.

This also removes the CTEST_UPDATE_COMMANDS_OUTPUT_FILE file before it gets
created again to avoid printing an old file when CTEST_UPDATE_VERSION_ONLY=ON
(a very rare case but still confusing).

Added documentation for CTEST_UPDATE_VERSION_ONLY=ON.

Added test for CTEST_UPDATE_VERSION_ONLY=ON.

NOTE: I was not able to add a strong test to show that the Update.txt and the
ctest_update() commands file are not written.  That would be a harder thing to
check for in the output.