Set CTEST_CHANGE_ID (#154)
It is not really possible to add a test for this.  It is just a var you set
and then CTest adds that to an XML file and that gets uploaded to CDash.
There really is nothing to test locally.

I just added reading in from env and documentation (with more detail that what
is provided by offical CDash documentation).