Change internal TriBITS target_link_libraries() to PUBLIC (#299)
This should fix the propagation of Kokkos compiler flags added on Kokkos
library targets using the INTERFACE_COMPILE_OPTIONS target property.  Because
Kokkos uses target_compile_options( ... PUBLIC ... ) and
target_link_libraries( ... PUBLIC ... ) and TriBITS was using old-style
target_link_libraries() with no PUBLIC/PRIVATE/INTERFACE, it was likely
defaulting to PRIVATE so the INTERFACE_COMPILE_OPTIONS target property would
not get propagated along.  This behavior does not seem to be documented in
official CMake documentation but it is documented in the book "Professional
CMake 5th Edition" which says:

> In particular, if a project defines a chain of library dependencies with a
> mix of both old and new forms of the command, the old-style form will
> generally be treated as PRIVATE.

I have not tested this yet but I strongly suspect this will fix the problem of
Trilinos customers linking to imported Trilinos CMake targets and not getting
the Kokkos compiler options directly from the targets.