WIP PASSED: Use add_library( ... INTERFACE ...) for external lib and explicit upstream lib files (#299)
Here the link actually works!  The trick is that you list the explict library
paths of the upstream libraries in target_link_libraries( ... INTERFACE ... )
in order instead of the targets.  That seems to trick/fool (or whatever) CMake
into putting those libs on the link link in the correct order.  It lists the
same libs more than once on the link line but at least it gets the job done.
This is a hack but what else can you do?