WIP FAILED: Try using -l<lib> -L<dir> for IMPORTED_LOCATION (#299)
This allows you to fix the ordering dependency of the libraries by being able
to set target_link_libraries() dependencies correctly but it does not allow
you to build because CMake puts in dependency rules like:

  packages/wrap_external/CMakeFiles/WrapExternal_libs: \
    -lexternal_func -L<build-dir>/packages/wrap_external/external_func

which fails the build because those are not files.  If you comment those
makefile dependency lines out of two generated 'build.make' files (just grep
for "[-]lexternal_func"), then you can get past those build errors.

Then the build fails with the link of the executable because CMake generates
the link line as:

  <compiler> <options> <objects> -o <out> \
    "-lexternal_func -L<build-dir>/packages/wrap_external/external_func" \
    ../with_subpackages/a/libpws_a.a ../simple_cxx/src/libsimplecxx.a

That fails the link because of the quotes around:

    "-lexternal_func -L<build-dir>/packages/wrap_external/external_func"

If you remove those quotes by manually modifying the generated 'link.txt' file
to give:

  <compiler> <options> <objects> -o <out> \
    -lexternal_func -L<build-dir>/packages/wrap_external/external_func \
    ../with_subpackages/a/libpws_a.a ../simple_cxx/src/libsimplecxx.a

then the build works perfectly with all of the arguments in the correct order
on the link line!

So what CMake needs is an IMPORTED library where you can set a property called
something like IMPORTED_LIBRARY_LINK_OPTIONS that is used instead of
IMPORTED_LOCATION.  Otherwise, I don't know how I can make this work for the
existing ATDM Trilinos builds where some libraries are specified as -l<lib>,
-L<dir>, and other link options that have to be in a very specific order.