WIP PASSED: Convert -l<lib> and -L<dir> to IMPORTED_LIBNAME and INTERFACE_LINK_OPTIONS (#299)
This commit converts "-L<dir> -l<lib>" to IMPORTED_LIBNAME and
INTERFACE_LINK_OPTIONS properties for the imported 'external_func' library.
It shows that you can take a link line with -l<lib> and -L<dir> options and
use that to build a link like that has -L<dir> and -l<lib> with the -l<lib>
arguments in the right order.  For example, in this case it build the link
line:

  <compiler> <options>\
    -L<build-dir>/packages/wrap_external/external_func \
    packages/wrap_external/<subdirs>/run_external_func.cpp.o \
    -o packages/wrap_external/WrapExternal_run_external_func.exe \
    -lexternal_func \
    packages/with_subpackages/a/libpws_a.a \
    packages/simple_cxx/src/libsimplecxx.a

This the IMORTED_LINK_OPTIONS are listed before any of the object or library
files but the -lexternal_func is listed in the right location in relation to
its upstream dependencies.

NOTE: Exporting of the exteranl_func target is not allowed in this case so the
WrapExternal package must be disabled so I put that logic back in.

NOTE: For Some reason, I also had to adjust a test checking for makefile
ouptut that changes produced by CMake.