Add isme() functions to HeaderOnlyTpl and SimpleTpl and use in downstream (#299)
* Added isme() functions to HeaderOnlyTpl and SimpleTpl TPLs

* Used the functions HeaderOnlyTpl::isme() and SimpleTpl::isme() in
  SimpleCxx::deps()

* Updated TribitsExampleApp and TriBITS tests using them to adjust.

* Updated TribitsExampleProject tests to adjust.  (Amazingly, there are no
  runings of the native TribitsExampleProject test suite that enable SimpleTpl
  and therefore I just hard-coded the deps for SimpleCxx as 'headeronlytpl'
  for now.  I will have to update all of these once I run the test suite with
  SimpleTpl enabled.)

NOTE: This required adding -DBUILD_SHARED_LIBS=ON to the build of SimpleTpl
when building shared libs for TribitsExampleApp.  Before, there must have not
been any real linking going on and the link failed without adding this option.
This shows we needed this extra code to make this a proper test of linking.