Split up test/core/ExamplesUnitTests/CMakeLists.txt (#299)
This just splits up the tests into individual files and pulls them in with
includes.  But I think we can do better to separate these sets of tests but
putting them into separate subdirs so they can't interact with each other by
accident but we can work on that later.  At least this reduces the size these
files and makes finding the different sets of tests easier.