Remove support for export makefiles (#299)
This removes all of the code, test, and documentation that I could related to
export makefiles that generate Makefile.export.* files.

NOTE: All of the TriBITS tests do *not* pass because the
TriBITS_CTestDriver_TribitsExMetaProj_* tests pull down old versions.  These
repos will need to be updated and then update the tests for those tests to
pass.