Update TriBITS_CTestDriver_TribitsExMetaProj_* tests after removing export makefiles (#299)
These tests are updated after I snapshotted the three example repos into the
'tribits-github-snapshot' branches, merged to the 'master' branches, then
ff-merged to the 'for-testing' branches, and pushed up to GitHub.  Here I had
to update the various git commit SHA1 and other info for this change.
Updating the snapshots fixed the build problems related to removing support
for export makefiles.