Allow libexternal_func.a or WrapExternal_run_external_func.exe as most recent modifed (#299)
It seems that even though TEST_14 output clearly shows that libexternal_func.a
gets built before WrapExternal_run_external_func.exe showing:

  [ 85%] Generating external_func/libexternal_func.a
  [ 85%] Built target build_external_func
  [100%] Built target WrapExternal_run_external_func

the code the returns the most recent modified binary file can return
libexternal_func.a instead of WrapExternal_run_external_func.exe because they
have exactly the same timestamp (down to the second) because they build so
quickly.

This change updates the criteria for the test to allow either of these files
to be returned as the most recently modified binary files.

In a real use case, all that matters is that some binary file show up as the
most recent modified file and therefore does not require a complete rebuild of
the package (blowing away the build dir and starting from scratch).