WIP PASSED: Use separate IMPORTED INTERFACE lib and INTERFACE lib to handle -l -L options (#299)
This is a successful experiment for the approach where a separate IMPORTED
INTERFACE library target is used for each -l<lib> argument and then an
INTERFACE library target is used to link against these and set the -L<dir>
options using target_link_options().  This is only a single set of -l and -L
options but the technique is demonstrated and is clear how to generalize to
more -l and -L arguments.

In addition, I also realized that the previous code was not bringing in the
optional MixedLang library correctly (which shows that external_func is not
using it the MixedLang library, which I need to fix at some point).

The updated link line is:

  <compiler> <options> \
    -L<build-dir>/packages/wrap_external/external_func \
    packages/wrap_external/<subdirs>/run_external_func.cpp.o \
    -o packages/wrap_external/WrapExternal_run_external_func.exe \
    packages/with_subpackages/a/libpws_a.a \
    -lexternal_func \
    packages/with_subpackages/a/libpws_a.a \
    packages/simple_cxx/src/libsimplecxx.a \
    packages/mixed_lang/src/libmixedlang.a \
    -lmpifort  -lgfortran  -lquadmath

That is a correct link line.  I am not sure why libpws_a.a shows up twice on
the link line but that may be due to the way that TriBITS is setting link
dependencies automatically.  (TriBITS needs to change to manual linking of
targets inside of a package to fix this.)

I think this is it!