WIP: Add <Package>::all_libs target (#299)
I have not actually used this target but it seems to do the right thing,
mostly.  It will need to be fixed up but it is close.

I also added a function tribits_get_all_build_targets_including_in_subdirs()
that may be handy in the future.  This can replace a function I wrote for
build-stats in Trilinos.