WIP: Use NAMESPACE:: for targets in <Package>Config.cmake files (#299)
This changes to use <Package>:: namespaced targets for libraries created in
TriBITS packages.  Since this also supports installation testing that reads in
<Package>Config.cmake files, this has to also change over the
${PACKAGE_NAME}_LIBRARIES variable to store a list of namespaced library
targets.  This maintains perfect backward compatiblity as long as clients are
they are just linking against ${${PACKAGE_NAME}_LIBRARIES} as now the imported
targets are namespaced as well.

Very nice.