Write out <tplName>ConfigVersion.cmake and disallow finding <tplName>Config.cmake (#299, #427)
This is one solution to the problem of finding <tplName>Config.cmake files in
the CMAKE_INSTALL_PREFIX directory in the middile of a FindTpl<tplName>.cmake
module.  If you have to write a FindTpl<tplName>.cmake module, you are not
expecting/wanting to find a TriBTIS-written <tplName>Config.cmake file.  See
the comment in TribitsProcessEnabledTpl.cmake for more details.

This should fix the bug reported in #427 where the find_package(HDF5 ...) call
in FindTPLHDF5.cmake was finding HDF5Config.cmake in the install tree written
by TriBITS in a previous installation.

I am not sure this is the best solution to this problem but it has the
advantage that we keep the <Package>Config.cmake files for external package
and TriBITS package side-by-side.  I suspect that I will implement another
solution but at least this is a start to writing <tplName>ConfigVersion.cmake
files.