Add mode to create specialized Tpl1Config.cmake file with find_dependency() (#299)
As part of this, I had to move the logic to write the default version file
<tplName>ConfigVersion.cmake to the calling function
tribits_process_enabled_tpl() and moved the addition of the install targets
there as well.  Therefore, those commands will get called reguardless if
tribits_tpl_find_include_dirs_and_libraries() gets called in the
FindTPL<tplName>.cmake module or not.

I also added a test for this case as well of course.

But note that there are not tests that actually use the
TplConfig[Version].cmake files yet.  That will come next.