WIP: Porting to work on native Windows (#299)
I have gotten the TribitsExampleApp tests to build on native Windows with MS
VC++ but I can't get the final app.exe executable to run.  It can't find the
DLLs.  I tried prepending the PATH with the ctest ENVIRONMENT argument but I
can't get that to work.  When I manaully prepend PATH in the dos shell, I can
get it to run.

I will try to manually copy the dll files in the next commit for WIN32 and see
what that does but I want to record this attempt.

NOTE: To get the test binaries to build, I shortened the test names (and
therefore also the and directory names) a little to get the total paths under
260 chars (which seems to be a hard limit in MS VC++).  I also moved my build
dir to a shorter base dir.