Fix failing test for <ParentPackage>Config.cmake missing subpackage includes (#299)
This sets <ParentPackage>_ENABLE_<SubPackage>=ON if the subpackage is enabled
even if optional packages are disabled.  This will fix trilinos/Trilinos#9972
and trilinos/Trilinos#9973.

This also updates the logic that generates <Package>Config.cmake files to only
include <UpstreamPackage>Config.cmake files for direct dependencies, not all
dependencies.  (The indirect includes should take care of the rest.)