Refactor to use CMAKE_CURRENT_FUNCTION to eliminate duplication (#299)
Now when you copy and paste one of these large test functions, you just need
to change the name of the function and that is it.

NOTE: I had to remove '_test' from the name of the test function so that that
name could be used in the CTest test name.