Handle/allow arbitrary linker arguments in TPL_<tplName>_LIBRARIES (#299, #433)
Turns out some Trilinos configure scripts (including some of the ATDM Trilinos
configure scripts) use arbitary linker arguments like '-mkl' that are not -L
or -l.  These arguments need to be moved forward on the link link in the
'all_libs' target.  Since this is not very general, I added a 'NOTE' printed
to STDOUT in order to warn about this.

NOTE: This now allows an agument like: '-o some-name' and it will move '-o'
forward and in the next commit the bare 'some-name' arg will be interepreted
as a library name and be handled the same as '-l<libname>'.  But obviously
splitting up the arguments `-o some-name` by moving '-o' forward by itself and
treating `some-name` as a link library is just wrong but I don't see how we
can avoid that.