Add util to TribitsSimpleExampleApp and other stuff (#299)
This adds a 'util' program that only uses the libs in the WithSubpackagesB
package.  This shows how to select just a subset of libraries from a package
read in after calling find_package(<ParentProject> ...).

This now makes the TribitsSimpleExampleApp and TribitsOldSimpleExampleApp
projects almost identical where the former uses modern CMake and the modern
TriBITS-generated targets and the latter works with old and new versions of
TriBITS.

Also:

* Switched to add_test(NAME ... COMMAND ...)
* Updated test drivers for the new 'util' test