snapshot-dir.py: add support for git commit --no-verify (SPAR-614)
This allows you to create the commit and skip pre-commit checks if needed.

Motivation: SPARC uses clang-format in a pre-commit hook and when you are
using snapshot-dir.py to snapshot googletest into one of the SPARC repos, you
need to be able to turn off calling this pre-commit check so that you can
create the commit with the commit message produced by this script.  You can
then amend the commit message after the fact and create a new commit to
address formatting after that.