tribits_tpl_find_include_dirs_and_libraries(): <tplName>_LIBRARY_NAMES cache var triggers lib finds (#299)
It seems reasonable that if the user needs to define libraries for a TPL with
what otherwise is defined as a TPL with only header files (specified by the
FindTPL<tplName>.cmake file by setting 'REQUIRED_HEADERS' but not
'REQUIRED_LIBS_NAMES') then setting the cache var TPL_<tplName>_LIBRARY_NAMES
should trigger the find for specified libraries.

I will add a real test for this next.