Add support for <tplName>_ALLOW_PACKAGE_PREFIND (#299)
This is needed to force off the prefind for TPLs that are set up to use
find_package().

I needed to add this for the updated TribitsExampleProject2_find_tpl_parts
tests that only set CMAKE_PREFIX_PATH but still wanted to run the individual
find_file() and find_library() functions.