Remove include_guard() from package and project config files (#299)
As I noted before in prior commit "Add support for TPL_<tplName>_DEPENDENCIES
in generation of TPL <Package>Config.cmake files (#299)", include_guard() has
problems that make it unsafe to use.  Instead, you check to make sure you are
not trying to define targets that are already existing.  But resetting vars is
fine.