Switch to relying on TRIBITS_TESTONLY_LIB property only (#299)
This gets rid of logic that was looking for <lib>_INCLUDE_DIRS.  With modern
CMake, the include dirs are carried along with the library target object
itself so there is no need for this var (but we don't remove the var yet, that
comes in later commit).

As part of this, I created a new module TribitsLibIsTestOnly.cmake with
functions to reduce duplication.