WIP: Remove old TriBITS manual manipulation of includes and libs (#299)
There is still a lot of work to clean up all of the junk that is being done in
TribitsWriteClientExportFiles.cmake.  But this is a good start.

NOTE: This does **not** handle the extra link flags/libraries given in
<Project>_EXTRA_LINK_FLAGS that is collected into the library target
'last_lib'.  That is an open problem that I am going to have to solve in some
way :-(